+++ src/com/app/service/SingleBuy.java
@@ -3,6 +3,6 @@ package com.app.service;
 public interface SingleBuy {
 
 	public abstract void singleBuyProduct(int userID, int productID,
-			int addressID);
+			long addressID);
 
 }
